Do the /schedule redirect with a middleware
