+++ next.config.js
@@ -1,22 +1,7 @@
 // @ts-check
 /** @type {import('next').NextConfig} */
 
-const { fetchSchedule } = require("./lib/schedule");
 const live = require("./data/live.json");
-const format = require("date-fns/format");
-
-const getScheduleDays = async () => {
-  const schedule = await fetchSchedule();
-
-  const potentialUnsortedDays = Array.from(new Set(Object.keys(schedule.days)));
-
-  return potentialUnsortedDays.sort((a, b) => {
-    const aDate = new Date(a);
-    const bDate = new Date(b);
-
-    return aDate.getTime() - bDate.getTime();
-  });
-};
 
 const securityHeaders = [
   {
@@ -48,9 +33,7 @@ const securityHeaders = [
 const nextConfig = {
   reactStrictMode: true,
   async redirects() {
-    const days = await getScheduleDays();
-
-    const redirects = [
+    return [
       {
         source: "/cfp",
         destination: "https://program.europython.eu/europython-2022/cfp",
@@ -67,20 +50,6 @@ const nextConfig = {
         permanent: false,
       },
     ];
-
-    if (days.length > 0) {
-      const today = format(new Date(), "yyyy-MM-dd");
-
-      let day = days.includes(today) ? today : days[0];
-
-      redirects.push({
-        source: "/schedule",
-        destination: `/schedule/${day}`,
-        permanent: false,
-      });
-    }
-
-    return redirects;
   },
 
   async headers() {
