+++ package.json
@@ -15,7 +15,7 @@
     "date-fns-tz": "^1.3.5",
     "glob": "^8.0.3",
     "gray-matter": "^4.0.3",
-    "next": "12.0.10",
+    "next": "^12.2.2",
     "next-mdx-remote": "^3.0.8",
     "react": "17.0.2",
     "react-dom": "17.0.2",
